Fixed prettier not working when formatting with conform + fixed not being able to disable copilot using Copilot disable
Fixed formatting of conform: this is due to the error where overwriting
of the conform config formatting, this has thus been move to another
autocmd location.

Copilot disable issue can be solved with the copilot version, just need
to update this.

Note that do not update bufferline as the recent update causes it to not
be able to have multiple buffers.